Plataforma Agentes_v3/requirements.txt
flask==3.0.1
flask-login==0.6.3
flask-sqlalchemy==3.1.1
openai==1.8.0
requests
psycopg2
sqlalchemy==2.0.25
python-dotenv
PyPDF2==3.0.1
weasyprint
pyodbc
azure-storage-blob
doc2text
llama-index-core
llama-index-readers-file
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai